+++ translation_finder/discovery/files.py
@@ -47,6 +47,13 @@ class GettextDiscovery(BaseDiscovery):
             mono["file_format"] = "po-mono"
             yield mono
 
+    def fill_in_new_base(self, result):
+        super(GettextDiscovery, self).fill_in_new_base(result)
+        if "new_base" not in result:
+            pot_name = result["filemask"].replace("po/*/", "pot/") + "t"
+            if self.finder.has_file(pot_name):
+                result["new_base"] = pot_name
+
 
 class QtDiscovery(BaseDiscovery):
     """Qt Linguist files discovery."""